Discard/truncate only files that failed to save, version 0.2.2
If the received file could not be saved, it was discarded along with all
subsequent files in the current transfer. The end of the transfer is
detected by the expiration of the no-data (inter-character) timeout.

After this modification, only the file that failed to save is discarded/
truncated. The following files are not affected.