Add cmdline argument --storage-file-exists=, version 0.3.0
Defines what to do if a file with the given name already exists
in the storage.

drop     - drop the received file
replace  - replace the file in storage with the received one

replace is the default

If a received file is dropped or a file in the storage is replaced,
a WARNING is printed. In previous versions, silent replacement of a file
in the storage occurred.